+++ src/app/(pages)/(albaform)/alba/[formId]/components/FormActions.tsx
@@ -34,7 +34,7 @@ export default function FormActions({ formId, albaFormDetailData }: FormActionsP
     try {
       await axios.delete(`/api/forms/${formId}`);
       toast.success("성공적으로 삭제되었습니다.");
-      router.push(`/albalist`);
+      router.push(`/alba-list`);
     } catch (error) {
       console.error(error);
       toast.error("삭제 중 오류가 발생했습니다.");
