+++ src/app/(pages)/my-albaform/(role)/applicant/page.tsx
@@ -50,7 +50,7 @@ export default function ApplicantPage() {
       if (!user) {
         router.push("/login");
       } else if (user.role === userRoles.OWNER) {
-        router.push("/myalbaform/owner");
+        router.push("/my-albaform/owner");
       }
     }
   }, [user, isUserLoading, router]);
