+++ src/app/(pages)/mypage/components/sections/PostsSection.tsx
@@ -92,7 +92,7 @@ export default function PostsSection() {
                     {page.data.map((post) => (
                       <div key={post.id} className="flex justify-center">
                         <Link
-                          href={`/albatalk/${post.id}`}
+                          href={`/alba-talk/${post.id}`}
                           className="block cursor-pointer transition-transform duration-300 hover:scale-[1.02]"
                         >
                           <div className="rounded-lg border border-grayscale-100 bg-white p-6">
