+++ src/app/components/card/board/BoardComment.tsx
@@ -59,7 +59,7 @@ const BoardComment = ({ id, postId, postTitle, comment, updatedAt, isAuthor = fa
   const dropdownOptions = [
     {
       label: "게시글 보기",
-      onClick: () => router.push(`/albatalk/${postId}`),
+      onClick: () => router.push(`/alba-talk/${postId}`),
     },
     {
       label: "삭제하기",
