+++ src/app/components/card/board/CardBoard.tsx
@@ -73,7 +73,7 @@ const CardBoard = ({
   };
 
   const handleEditTalk = () => {
-    router.push(`/albatalk/${id}/edit`);
+    router.push(`/alba-talk/${id}/edit`);
   };
 
   const dropdownOptions = [
@@ -100,12 +100,12 @@ const CardBoard = ({
       className={`h-[220px] w-[320px] rounded-[16px] border p-4 lg:h-[240px] lg:w-[360px] ${
         variant === "primary" ? "border-primary-orange-100 bg-primary-orange-50" : "border-line-100 bg-grayscale-50"
       } cursor-pointer`}
-      onClick={() => router.push(`/albatalk/${id}`)}
+      onClick={() => router.push(`/alba-talk/${id}`)}
       role="button"
       tabIndex={0}
       onKeyDown={(e) => {
         if (e.key === "Enter" || e.key === " ") {
-          router.push(`/albatalk/${id}`);
+          router.push(`/alba-talk/${id}`);
         }
       }}
     >
