+++ src/app/components/layout/Header.tsx
@@ -88,14 +88,14 @@ export default function Header() {
           </Link>
 
           <div className={menuStyle}>
-            <Link href="/albalist" className={getLinkClassName("/albalist")}>
+            <Link href="/alba-list" className={getLinkClassName("/alba-list")}>
               알바 목록
             </Link>
-            <Link href="/albatalk" className={getLinkClassName("/albatalk")}>
+            <Link href="/alba-talk" className={getLinkClassName("/alba-talk")}>
               알바 토크
             </Link>
             {user && (
-              <Link href="/myalbaform" className={getLinkClassName("/myalbaform")}>
+              <Link href="/my-albaform" className={getLinkClassName("/my-albaform")}>
                 내 알바폼
               </Link>
             )}
